+++ blacklight_range_limit.gemspec
@@ -18,10 +18,8 @@ Gem::Specification.new do |s|
   s.license     = "Apache 2.0"
 
   s.add_dependency "rails", ">= 3.0", "< 5.0"
-  s.add_dependency "jquery-rails" # our JS needs jquery_rails
-  # for blacklight, we want to allow 5.0.0.preX, as well as all 5.x.y, 
-  # but not 6. can't seem to make it do so other than this:
-  s.add_dependency "blacklight", ">= 5.0.0.pre4", "< 6"
+  s.add_dependency "jquery-rails" # our JS needs jquery_rails  
+  s.add_dependency "blacklight", "~> 5.14"
 
   s.add_development_dependency "rspec", "~> 3.0"
   s.add_development_dependency "rspec-rails"
