+++ lib/blacklight_range_limit/view_helper_override.rb
@@ -103,6 +103,10 @@
 
       return array
     end
+
+    def range_config(solr_field)
+      BlacklightRangeLimit.range_config(blacklight_config, solr_field)
+    end
     
   end
